Move CLI config to './cli'
