+++ cli/src/contexts/cities-of-brazil/actions/update.js
@@ -3,7 +3,6 @@ import fs from "fs-extra";
 import cliProgress from "cli-progress";
 import {
   addDays,
-  endOfDay,
   isAfter,
   isBefore,
   parseISO,
@@ -50,7 +49,7 @@ import {
   POLYFILES_LEVEL_1_DIR,
   POLYFILES_LEVEL_2_DIR,
   POLYFILES_LEVEL_3_DIR,
-} from "../../config.js";
+} from "../config.js";
 
 // Set concurrency limit
 const limit = pLimit(20);
